+++ src/databricks/sql/backend/filters.py
@@ -19,7 +19,7 @@ from typing import (
 
 from databricks.sql.utils import JsonQueue, SeaResultSetQueueFactory
 from databricks.sql.backend.types import ExecuteResponse, CommandId
-from databricks.sql.backend.models.base import ResultData
+from databricks.sql.backend.sea.models.base import ResultData
 
 if TYPE_CHECKING:
     from databricks.sql.result_set import ResultSet, SeaResultSet
